Type Alias local_storage_initializer::ClsDataImage

pub type ClsDataImage = LocalStorageDataImage<Cls>;

Aliased Type

struct ClsDataImage { /* private fields */ }

Implementations

impl<T> LocalStorageDataImage<T>where T: LocalStorage,

pub const fn new() -> Self

Creates an empty data image.

pub fn inherit(&mut self, other: &Self)

Inherits the data from another data image.

Panics

Panics if the other image has a longer length, or (on x86_64) if the other image has a differently sized static area.

impl LocalStorageDataImage<Cls>

pub unsafe fn set_as_current_cls(&self)

Sets the data image.

Safety

The data image must not be dropped until another data image replaces it.